db Add Latest model to store the Latest dataset
Instead of storing the dataset for Latest (was Canonical) in a
combination of memory and a file cache store it in the database instead.
This has the advantage of querying and the lack of needing to
synchronise what is in the database and what is in the dataset.

Fixes issue #13