commands create_datagetter_data Flatten grant data output
Flatten the grantdata so that it contains the additional_data object at
the same level of the grant data. This is to maintain compatibility with
grantnav.
Also add add option to indent output with default to none in order that
we can save some disk space and processing time when we don't need the
output as human readable.

Related issue #20