poll_datastore Move fetch_data_package to AFTER poll file written
If we write the last poll date file there is a risk that a second
download will be started if it happens in the poll frequency. e.g.
download takes >5min and poll is every 5min.

Also switch subprocess to check_call rather than just call in order that
we capture any non 0 exit status of the load_grantnav_script