additional_data: Add geolookups model and source
This command fetches data from https://github.com/drkane/geo-lookups on LSOA,
MSOA, ward and LA. Each grant is checked in three ways to see if any of the
geocodes for those areas are used. In order of priority they are:

- beneficiaryLocation
- recipientOrganization -> Location
- recipientOrganizationLocation (from additional_data)

If locations are found at any level then the subsequent levels aren't checked.
It is possible to add more than one location though, if there were eg more than one beneficiaryLocation supplied.

The data that is added to additional_data allows for standardised local
authority and higher geographies. For LSOA, MSOA & Ward it also adds a lat and
longitude corresponding to the centre of the area.