db: load_datagetter_data: Fix exception handling for SourceFile objects
Remove the field "grants" from sourcefile objects. If a SourceFile has
failed to download or is invalid in some other way trying to set the
total number of grants will result in an exception. This exception was
being excepted causing the failed sources not to be created and
therefore stopping the Latest best generator from working correctly.

Remove from api as this total number of grants is now available in
the aggregate data anyway.