quality_data: Rework the aggregated to publisher quality data
Count the number of instances the that a publisher's source file doesn't
have the required badge and aggregate this to the publisher level.

Additional work to separate the calculations for the total number of grants to
orgs/individuals depending on the quality test.

Add some additional comments where we need to do more work.

Tests:
- Swap the assert round so that the output diff is correct direction.

Command:
- Fix the multiprocessing issue
- https://github.com/ThreeSixtyGiving/datastore/issues/126