db: commands: Add load_data_package
This command allows us to load a processed data package into the
datastore that has been created by the datastore itself. This skips the
lengthy additional_data stage.

The main use case for this is to be able to export data from the live
datastore into a development instance in a timely way.