tools: data_run: Remove MAX_TOTAL_RUNS_IN_DB if statement
The datastore now handles an argument for deciding the total number of
datagetter runs held in the database. Remove this so it doesn't
conflict/add another layer on top of the arguments supplied.

We leave MAX_TOTAL_RUNS_IN_DB in the script we still have:

```
MAX_PACKAGE_AGE_DAYS=`expr $MAX_TOTAL_RUNS_IN_DB + 2`
```