additional_data: geo_lookup: Always add all geo lookup data
To be able to use the locationLookup fields for best, beneficiary and
recipient organisation we need to add them all to our locationLookup data.

The "best location" is still the first one in the array of
locationLookup data. In future grantnav will index based on the source
as well so as to provide a selection as to which the user wants (best,
beneficiary, recipientOrganization).