db: Update entities atomically
The update_entities() function deletes all entities before creating new
ones. Without being contained in a transaction, this resulted in all the
entities dissapearing from the database for a time while the function
runs. This commit wraps the function in a transaction, so the old
entities data remains until the function completes.