db: Refactor Publisher entities
This commit refactors when and how Publisher entities are created.
Previously, Publisher objects were created at the same time as data was
loaded in, and related to the GetterRun they were created by.
This meant duplicate Publisher objects being created, one for each
GetterRun that was still kept in the DB, and led to confusion as to the
current state of Publishers within the live dataset, and propagated
sometimes out of date info to GrantNav.

This commit refacts Publishers to be created by `manage_entities_data`
along with Funders and Recipients, using SourceFile data, and no longer
related to individual GetterRuns. This means the set of Publisher
objects should now always reflect the current live dataset.